--- Northwind/FrontEnd/Controllers/CategoryController1.cs
@@ -1,81 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-
-namespace FrontEnd.Controllers;
-public class CategoryController1 : Controller
-	{
-	// GET: CategoryController1
-	public ActionResult Index()
-		{
-		return View();
-		}
-
-	// GET: CategoryController1/Details/5
-	public ActionResult Details(int id)
-		{
-		return View();
-		}
-
-	// GET: CategoryController1/Create
-	public ActionResult Create()
-		{
-		return View();
-		}
-
-	// POST: CategoryController1/Create
-	[HttpPost]
-	[ValidateAntiForgeryToken]
-	public ActionResult Create(IFormCollection collection)
-		{
-		try
-			{
-			return RedirectToAction(nameof(Index));
-			}
-		catch
-			{
-			return View();
-			}
-		}
-
-	// GET: CategoryController1/Edit/5
-	public ActionResult Edit(int id)
-		{
-		return View();
-		}
-
-	// POST: CategoryController1/Edit/5
-	[HttpPost]
-	[ValidateAntiForgeryToken]
-	public ActionResult Edit(int id, IFormCollection collection)
-		{
-		try
-			{
-			return RedirectToAction(nameof(Index));
-			}
-		catch
-			{
-			return View();
-			}
-		}
-
-	// GET: CategoryController1/Delete/5
-	public ActionResult Delete(int id)
-		{
-		return View();
-		}
-
-	// POST: CategoryController1/Delete/5
-	[HttpPost]
-	[ValidateAntiForgeryToken]
-	public ActionResult Delete(int id, IFormCollection collection)
-		{
-		try
-			{
-			return RedirectToAction(nameof(Index));
-			}
-		catch
-			{
-			return View();
-			}
-		}
-	}
